🚚 Move `dictionary` arbitraries into /arbitrary
