+++ src/arbitrary/_internals/mappers/KeyValuePairsToObject.ts
@@ -1,0 +1,9 @@
+/** @internal */
+export function keyValuePairsToObjectMapper<T>(items: [string, T][]): { [key: string]: T } {
+  // Equivalent to Object.fromEntries
+  const obj: { [key: string]: T } = {};
+  for (const keyValue of items) {
+    obj[keyValue[0]] = keyValue[1];
+  }
+  return obj;
+}
